requirements.txt
# Architecture Review Agent Dependencies

# Core Python libraries
argparse>=1.4.0  # Built-in, but specified for clarity

# Document processing dependencies
python-docx>=0.8.11      # For Word .docx files
python-docx2txt>=0.8      # For older Word .doc files
PyPDF2>=3.0.0            # For PDF processing
pdfplumber>=0.9.0        # Better PDF text extraction
openpyxl>=3.0.0          # Excel export support
beautifulsoup4>=4.11.0   # HTML processing
lxml>=4.9.0              # XML/HTML parser for BeautifulSoup

# Optional dependencies for enhanced functionality
pyyaml>=6.0      # For YAML configuration support
requests>=2.28.0  # For fetching remote standards/templates
jinja2>=3.1.0    # For advanced report templating
click>=8.0.0     # Alternative CLI framework
rich>=13.0.0     # Enhanced console output
tabulate>=0.9.0  # Table formatting for reports

# Enhanced functionality dependencies
pyyaml>=6.0.0     # YAML configuration support (required for config manager)
hashlib           # For session IDs (built-in)
dataclasses       # For enhanced data structures (built-in in Python 3.7+)
typing            # Type hints (built-in in Python 3.5+)
datetime          # Date/time handling (built-in)

# Optional enhanced features
pandas>=1.3.0     # Advanced data analysis (optional)

# Development dependencies (optional)
pytest>=7.0.0     # Testing framework
black>=22.0.0     # Code formatting
flake8>=5.0.0     # Code linting
mypy>=0.991       # Type checking
coverage>=6.0.0   # Code coverage reporting
